requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Marshmallow
marshmallow
marshmallow-sqlalchemy
psycopg2-binary
python-dotenv
flask-cors
gunicorn